Made the input file generic for use with Bluetooth controller or 360 controller
Setting up input files in Unity is always a huge pain. I have 3 good
input files now. One for the bluetooth controller, one for the 360
controller, and one that is generic. I should be able to use the generic
one to allow players to set up their own control systems.